Only clear queue if member has manage message perm

+++ commands/ClearQueueCommand.js
@@ -5,6 +5,7 @@ module.exports = {
   description: 'Clears music queue',
   guildOnly: true,
   async execute (message) {
+    if (!message.member.hasPermission('MANAGE_MESSAGES')) return message.channel.send('You cannot run this command!')
     db.query(`DELETE FROM music_queue WHERE guild = ${message.guild.id};`).catch(e => {
       console.error(e.stack)
       return message.channel.send(e)
